+++ src/maths/Maths.hpp
@@ -3,14 +3,14 @@
 #include <cmath>
 
 namespace Maths {
-  double NRoot(const double v, const double n) {
-    double out = v;
-    const double exp = 1. / n;
+  long double NRoot(const long double v, const long double n) {
+    long double out = v;
+    const long double exp = 1. / n;
     if (std::fmod(n, 1.) == 0.) {
       out = std::pow(out, exp);
     }
     else {
-      double absroot = std::pow(std::abs(out), exp);
+      long double absroot = std::pow(std::abs(out), exp);
       if (out < 0.) absroot *= -1.;
       out = absroot;
     }
